Move ante handler to auth module
The functionality required to support Textual signing is in the auth
module. This functionality is required by the ante handler which was
previously in the gears module. The gears module cannot depend on the
auth module since the auth module already depends on the gears module.
Moving the ante handler to the auth module fixes this issue and is where
the Cosmos SDK has the ante handler.