CBOR: Use BTreeMap instead of HashMap and OrderedHashMap
For determinstic serialization of maps, the iteration order must be
deterministic. This rules out using a HashMap. We could have used the
OrderedHashMap throughout, however we chose the BTreeMap since it's part
of the standard library.

The CBOR spec specifies that the order is given by the lexicgraphical
ordering of the CBOR serialized keys. ciborium provides the
CanonicalValue type for this purpose. Consequently we implement CBOR
serialization for BTreeMap's with CanonicalValue key type.