Change how textual rendering gets metadata
Previously the textual rendering sign_bytes_get method would get denom
metadata through the Context. This change means it now gets it through a
closure passed as an argument to the method. Ultimately the denom
metadata is controlled by the bank keeper which the Context has no
acess to. Which is why this changes the access to a closure.