Replace forked Decimal256 with CosmWasm import
There are a couple of methods which need to be added to CosmWasm's
Decimal256 type. Previously this was done by copy pasting the
implementation into this crate and manually adding the additional
methods. This change adds the additional methods by defining them inside
a trait then implementing the trait on the library Decimal256 type.