Update PrimitiveValueRenderer
Update the PrimitiveValueRenderer to return a Content rather than a
String. This removes the need for consumers to wrap the String in a
Content and handle any errors.

The PrimitiveValueRenderer is also effectively split into three traits,
corresponding to three different methods, format, try_format and
try_format_with_metadata. This allows the appropriate trait to be chose
when implementing a PrimitiveValueRenderer for a given type.

The types on which the DefaultPrimitiveRenderer are implemented is
updated to match those in the Cosmos SDK spec:
https://docs.cosmos.network/v0.50/build/architecture/adr-050-sign-mode-textual.