Make get_metadata in textual rendering fallible
The recently added ledger signing code uses textual rendering. However
the implementation of textual rendering's get_metadata handler on the
ledger side makes an rpc call to a node to get the denom metadata. This
rpc call is fallible however the get_metadata method expected all calls
for metadata to be succesful.