Replace AppError with anyhow::Error in ABCI handle
The fallible ABCIHandler trait methods returned an AppError, which is an
enum with many variants. However, since Gears doesn't care which variant
is returned it doesn't make sense to use AppError. All Gears cares about
is that the error can be converted to a String. anyhow::Error satisfies
this requirement and is onvenient to use for app developers.